pyproject.toml
[tool.poetry]
name = "tap-s3"
version = "0.1.0"
description = ""
authors = ["Michael Cooper <macoop2363@gmail.com>"]
packages = [
    { include = "tap_s3" }
]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.16.59"
python-dotenv = "^0.15.0"
singer-python = "^5.9.1"
pandas = "^1.2.1"
pyjanitor = "^0.20.10"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
moto = {extras = ["s3"], version = "^1.3.16"}
coverage = "^5.4"

[tool.poetry.scripts]
tap-s3 = 'tap_s3:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
